requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
binaryornot==0.4.4
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cookiecutter==2.1.1
Django==3.2.18
django-browser-reload==1.7.0
django-tailwind==2.2.2
dnspython==2.3.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.2.0
loguru==0.6.0
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.21.6
openai==0.27.2
pinecone-client==2.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-slugify==8.0.1
PyYAML==6.0
requests==2.28.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
threadpoolctl==3.1.0
tqdm==4.65.0
typing-extensions==4.5.0
tzdata==2023.2
urllib3==1.26.15
whitenoise==6.4.0
win32-setctime==1.1.0
yarl==1.8.2